moved file and directory variables to nsfrb.config
