+++ legacy/make_fringe_stop_tables.py
