+++ nsfrb/noise.py
@@ -4,6 +4,10 @@ import sys
 import os
 import glob
 
+from nsfrb.config import cwd,cand_dir,frame_dir,psf_dir,img_dir,vis_dir,raw_cand_dir,backup_cand_dir,final_cand_dir,inject_dir,training_dir,noise_dir,imgpath,coordfile,output_file,processfile,timelogfile,cutterfile,pipestatusfile,searchflagsfile,run_file,processfile,cutterfile,cuttertaskfile,flagfile,error_file,inject_file,recover_file,binary_file
+
+from nsfrb.config import *
+"""
 #f = open("../metadata.txt","r")
 #cwd = f.read()[:-1]
 #f.close()
@@ -15,7 +19,7 @@ from nsfrb.config import *
 #noise directory
 noise_dir = cwd + "-noise/" 
 output_file = cwd + "-logfiles/search_log.txt"
-
+"""
 
 def noise_update(noise,gridsize_RA,gridsize_DEC,DM,width,noise_dir=noise_dir,output_file=output_file):
     """
