--- process_server/process_flags.txt
@@ -1,1 +1,0 @@
-8
