+++ simulations_and_classifications/generate_source_images.py
@@ -7,12 +7,12 @@ import numpy as np
 import matplotlib.pyplot as plt
 import pandas as pd
 from tqdm import tqdm  
-from nsfrb.simulating import compute_uvw, add_complex_gaussian_noise, get_core_coordinates, apply_spectral_index
+from nsfrb.simulating import compute_uvw, add_complex_gaussian_noise, get_core_coordinates, apply_spectral_index, apply_phase_shift
 from nsfrb.imaging import uniform_image
 from nsfrb.config import NUM_CHANNELS, CH0, CH_WIDTH, AVERAGING_FACTOR, IMAGE_SIZE, c
 
 
-def generate_src_images(dataset_dir, num_observations, noise_std_low, noise_std_high, exclude_antenna_percentage, HA_low, HA_high, Dec_low, Dec_high, spectral_index_low, spectral_index_high, zoom_pix, tonumpy,inflate=False,noise_only=False,N_NOISE=1):
+def generate_src_images(dataset_dir, num_observations, noise_std_low, noise_std_high, exclude_antenna_percentage, HA_point, HA_source, Dec_point, Dec_source, spectral_index_low, spectral_index_high, zoom_pix, tonumpy,inflate=False,noise_only=False,N_NOISE=1):
     """
     This function generates images of sources observed with DSA-110 core antennas.
     It takes various parameters such as the dataset directory, the number of observations, 
@@ -26,10 +26,10 @@ def generate_src_images(dataset_dir, num_observations, noise_std_low, noise_std_
     - noise_std_low (float): Lower bound for noise standard deviation.
     - noise_std_high (float): Upper bound for noise standard deviation.
     - exclude_antenna_percentage (list[float]): Lower and upper bounds for excluded antenna percentage.
-    - HA_low (float): Lower bound for HA (Hour Angle).
-    - HA_high (float): Upper bound for HA (Hour Angle).
-    - Dec_low (float): Lower bound for Dec (Declination).
-    - Dec_high (float): Upper bound for Dec (Declination).
+    - HA_point (float): HA for center of image (Hour Angle).
+    - HA_source (float): HA of source (Hour Angle).
+    - Dec_point (float): DEC for center of image (Declination).
+    - Dec_source (float): DEC of source (Declination).
     - spectral_index_low (float): Lower bound for spectral index.
     - spectral_index_high (float): Upper bound for spectral index.
     - zoom_pix (int): Number of pixels to zoom in.
@@ -53,8 +53,10 @@ def generate_src_images(dataset_dir, num_observations, noise_std_low, noise_std_
     for obs in tqdm(range(num_observations), desc="Progress"):
         desired_shift_pixels_x = 0
         desired_shift_pixels_y = 0
-        HA = np.random.uniform(low=HA_low, high=HA_high)
-        Dec = np.random.uniform(low=Dec_low, high=Dec_high)
+        ##HA = np.random.uniform(low=HA_low, high=HA_high)
+        ##Dec = np.random.uniform(low=Dec_low, high=Dec_high)
+
+        
 
         # Calculate u_shift and v_shift in wavelengths
         #u_shift = desired_shift_pixels_x * pixel_resolution
@@ -66,11 +68,13 @@ def generate_src_images(dataset_dir, num_observations, noise_std_low, noise_std_
 
         exclude_antennas = np.random.choice(range(ANTENNA_COUNT), size=int(ANTENNA_COUNT * EXCLUDE_ANTENNA_PERCENTAGE), replace=False)
 
-        u_core, v_core, w_core = compute_uvw(x_core, y_core, z_core, HA, Dec) #core antennas
+        u_core, v_core, w_core = compute_uvw(x_core, y_core, z_core, HA_point, Dec_point) #core antennas -- this will make the PSF show up at the center
+        u_core_s, v_core_s, w_core_s = compute_uvw(x_core, y_core, z_core, HA_source, Dec_source) #use this to offset to the source
 
         V = [np.ones(len(u_core)) + 1j*np.ones(len(v_core)) for _ in range(NUM_CHANNELS)]
 
         u_core_new, v_core_new, w_core_new, V_new = [], [], [], []
+        u_core_s_new, v_core_s_new, w_core_s_new = [], [], []
 
         for i in range(ANTENNA_COUNT):
             for j in range(i+1, ANTENNA_COUNT):
@@ -79,10 +83,15 @@ def generate_src_images(dataset_dir, num_observations, noise_std_low, noise_std_
                     u_core_new.append(u_core[index])
                     v_core_new.append(v_core[index])
                     w_core_new.append(w_core[index])
+                    
+                    u_core_s_new.append(u_core_s[index])
+                    v_core_s_new.append(v_core_s[index])
+                    w_core_s_new.append(w_core_s[index])
                     V_new.append([v[index] for v in V])
 
         # Convert lists back to numpy arrays or appropriate data structures
         u_core, v_core, w_core = np.array(u_core_new), np.array(v_core_new), np.array(w_core_new)
+        u_core_s, v_core_s, w_core_s = np.array(u_core_s_new), np.array(v_core_s_new), np.array(w_core_s_new)
         V = np.array(V_new).T 
 
         reference_frequency_MHz = np.random.uniform(CH0, CH0 + NUM_CHANNELS * CH_WIDTH)
@@ -109,8 +118,14 @@ def generate_src_images(dataset_dir, num_observations, noise_std_low, noise_std_
                 wavelength = c / (avg_freq * 1e6)
                 chunk_u_core = u_core / wavelength
                 chunk_v_core = v_core / wavelength
+                chunk_u_s_core = u_core_s / wavelength
+                chunk_v_s_core = v_core_s / wavelength
 
                 # Apply the phase shift to chunk_V
+                #print("PHASE SHIFT APPLIED:",HA_source-HA_point,Dec_source-Dec_point)
+                chunk_V = [apply_phase_shift(v, chunk_u_core, chunk_v_core, HA_source-HA_point, Dec_source-Dec_point) for v in chunk_V]
+            
+            
                 #chunk_V_shifted = [apply_phase_shift(v, u_shift_rad, v_shift_rad) for v in chunk_V]
                 if inflate:
                     dirty_img = uniform_image(chunk_V, chunk_u_core, chunk_v_core, zoom_pix*2)
@@ -141,13 +156,13 @@ def generate_src_images(dataset_dir, num_observations, noise_std_low, noise_std_
             if tonumpy and N_NOISE > 1:
                 dirty_img_all_noise[:,:,k_n,:] = dirty_img_all
             if tonumpy:
-                np.save(os.path.join(observation_dir,f'final_img_{HA:.2f}_hr_{Dec:.2f}_deg.npy'),dirty_img_all)
+                np.save(os.path.join(observation_dir,f'final_img_{HA_point:.2f}_hr_{Dec_point:.2f}_deg.npy'),dirty_img_all)
             metadata.append({
                 #'filename': filename,
                 'desired_shift_pixels_x': desired_shift_pixels_x,
                 'desired_shift_pixels_y': desired_shift_pixels_y,
-                'HA': HA,
-                'Dec': Dec,
+                'HA': HA_point,
+                'Dec': Dec_point,
                 'Noise': noise,
                 'Spectral index': spectral_index,
             })
@@ -169,10 +184,10 @@ if __name__ == "__main__":
     parser.add_argument('--noise_std_low', type=float, default=1, help='Lower bound for noise standard deviation')
     parser.add_argument('--noise_std_high', type=float, default=2, help='Upper bound for noise standard deviation')
     parser.add_argument('--exclude_antenna_percentage', type=float, nargs=2, default=[0, 0.15], help='Lower and upper bounds for excluded antenna percentage')
-    parser.add_argument('--HA_low', type=float, default=0, help='Lower bound for HA')
-    parser.add_argument('--HA_high', type=float, default=10/180*np.pi, help='Upper bound for HA')
-    parser.add_argument('--Dec_low', type=float, default=-np.pi/2, help='Lower bound for Dec')
-    parser.add_argument('--Dec_high', type=float, default=np.pi/2, help='Upper bound for Dec')
+    parser.add_argument('--HA_point', type=float, default=0, help='Lower bound for HA')
+    parser.add_argument('--HA_source', type=float, default=0, help='Upper bound for HA')
+    parser.add_argument('--Dec_point', type=float, default=0, help='Lower bound for Dec')
+    parser.add_argument('--Dec_source', type=float, default=0, help='Upper bound for Dec')
     parser.add_argument('--spectral_index_low', type=float, default=-2, help='Lower bound for spectral index')
     parser.add_argument('--spectral_index_high', type=float, default=2, help='Upper bound for spectral index')
     parser.add_argument('--zoom_pix', type=int, default=25, help='Number of pixels to zoom in')
@@ -181,4 +196,4 @@ if __name__ == "__main__":
     parser.add_argument('--NNOISE',type=int,default=1,help='Number of noise realizations to generate; only 1 is saved as observation dataset')
     args = parser.parse_args()
 
-    generate_src_images(args.dataset_dir, args.num_observations, args.noise_std_low, args.noise_std_high, args.exclude_antenna_percentage, args.HA_low, args.HA_high, args.Dec_low, args.Dec_high, args.spectral_index_low, args.spectral_index_high, args.zoom_pix,args.tonumpy,args.inflate,args.NNOISE)
+    generate_src_images(args.dataset_dir, args.num_observations, args.noise_std_low, args.noise_std_high, args.exclude_antenna_percentage, args.HA_point, args.HA_source, args.Dec_point, args.Dec_source, args.spectral_index_low, args.spectral_index_high, args.zoom_pix,args.tonumpy,args.inflate,args.NNOISE)
